+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changes
 - Update invoice items to current collmex field names.
   (`#27 <https://github.com/gocept/gocept.collmex/pull/27>`_)
 
+- Add CustomerAgreement (CMXCAG)
+
 
 2.0.1 (2023-08-23)
 ------------------